# lib/api-zod/src/generated/types/companyListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CareerRadar API — personal placement OS for CS students
 * OpenAPI spec version: 0.2.0
 */
import type { Company } from './company';
import type { PaginationMeta } from './paginationMeta';

export interface CompanyListResponse {
  data: Company[];
  meta: PaginationMeta;
}